Data Deduplication: Breaking the Speed Limit of Digital Forensic Investigations
Abstract

Digital forensics is a relatively new field in both industry and
academia. Many digital forensic tools have been created in this short
time but the tools are never sufficient enough to deal with the
exponential backlog of cases digital investigators face. The backlog of
cases exists for a combination of reasons but two of the most prevalent
are, a lack of standardisation of digital forensic investigations and
the ease of public access to massive volumes of storage.

The research question proposed is, can the digital forensic
investigation process be quicker? The data deduplication tool aims to
speed up digital forensic investigations and offer digital forensics as
a service (DFaaS) to digital investigators. The concept behind the tool
is for a database in the cloud to store known acquired files where
digital investigators can compare files for their ongoing
investigations. The tool aims to speed up the process of disk
acquisition by comparing a file previously acquired to a newly acquired
file by the hash values from each file. If the hash values match it
means the database already has the file but if the hash values do not
match the file is then sent to the known files collection in the
database for future investigations.

The literature review chapter gives a detailed overview of some of the
key areas related to the data deduplication tool. The areas reviewed
include remote evidence acquisition, the current backlog, hard drive
forensics, information retrieval, DFaaS and other related work. The
literature review lays the foundation for what the data deduplication
tool is built upon. The design and flow of events of the tool are shown
in Chapter 3 with an overview of why Python and MongoDB were chosen as
the technologies to build the data deduplication tool.

The experimentation and results chapter consists of six tests carried
out with the data deduplication tool. The initial test is carried out
on Window 7,8,10 and Ubuntu to show the speed of reading the disk and
acquiring file metadata only. The following five tests run the data
deduplication tool from start to finish and are carried out on Windows
7, 8 and 10. The tests aim to prove how the implementation of the data
deduplication tool can speed up a digital forensic acquisition which in
turn will speed up a digital forensic investigation.
The body of work in this thesis outlines the real problem of digital
evidence backlogs world- wide. The data deduplication tool has been
built and rigorously tested with the primary aim of tackling this
problem. The literature review and design chapter explain the data
deduplication tool in detail and the tests carried out in Chapter 5
show the potential advantages of using the data deduplication tool in
industry.